Split `Card` related models into other files
